Add explicit access-level modifier support for imports in code generator (#2010)
## Motivation
This is a follow-up of https://github.com/grpc/grpc-swift/pull/2003.
As of 5.9, Swift supports access-level modifiers on imports. In an
upcoming Swift 6.x release, the default modifier will become `internal`.

## Modifications
This PR adds explicit access-level modifiers to the generated code and a
few other modules that use this generated code.

## Result
Explicit access level imports present in more places.