Fix casing of generated services and methods
Motivation:

Given a method name, say `ImportCSV`, from a `.proto` file, the generated method name in upper and lower camel case is `ImportCsv` and `importCsv` respectively. The generated method name (which is already expected to be in upper camel case) in upper and lower camel case should be `ImportCSV` and `importCSV` respectively.

Modifications:

- Replace the method used for generating service and method names in lower camel case with a newly implemented method.
- Do not convert the base names of services and methods to upper camel case as they are expected to already be in upper camel case.

Result:

The casing of service and method names will be preserved when generating stubs.