Simplify and update code gen tests (#14)
Motivation:

The protobuf code generator doesn't actually do a lot: it parses a
protobuf file descriptor and turns that into a request for the code
generator provided by the GRPCCodeGen module. This means that the tests
should focus on the conversion rather than the generated code (although
this is still useful to test at a high level).

As it stands all tests are done on the generated code rather than the
conversion to a code generator request. The existing tests also hand
roll a protobuf file descriptor which is tedious and error prone.

The upstream code generator also has changes which aren't yet reflected
here.

Modifications:

- Define some test proto files for which we generate serialized protobuf
file descriptor sets.
- Add a suite of code gen parser tests which check the request created
from the file descriptor proto.
- Add a single code generator test.
- Downgrade an access level from public to package as public wasn't
necessary.

Result:

Better tests