Generate code file for protos with no definitions
Motivation

To prepare for use in a SwiftPM build plugin which requires
deterministic output files and to match the behavior of
`proto-gen-swift` we should generate a source file even if no
definitions are found.

Modifications:

We no longer return early in the case of no definitions being found.

Result:

We will generate a Swift source file even if no definitions are found.