Generate code file for protos with no definitions (#27)
This change is broken out of
https://github.com/grpc/grpc-swift-protobuf/pull/26

### Motivation

To prepare for use in a SwiftPM build plugin which requires
deterministic output files and to match the behavior of
`proto-gen-swift` we should generate a source file even if no
definitions are found.

### Modifications:

* We no longer return early in the case of no definitions being found.
* Add empty proto file test
* Add a preamble to `foo-messages.proto`
* Don't unconditionally add a dependency on `GRPCProtobuf` if there are
no services
* Add the new test case code generation to `dev/protos/generate.sh`
* Depend on `grpc-swift` `main` to pick up empty file generation
changes.

### Result:

We will generate a Swift source file even if no definitions are found.