Use relative path for input files to protoc (#79)
Motivation:

protoc requires the import path (-I/--import-path) to be an exact prefix
of the input files. The plugin currently uses the absolute path of a
file for input paths without modifying the import path.

The result of this is that the user must always specify an absolute
import path (because the default chosen by is protoc is ".").

Modifications:

- Use the relative path for input files (this will be whatever the user
provides, which may be an absolute path) rather than the absolute path
which will get the full absolute path even if the user specified
"foo.proto" as input.

Result:

Users can do: swift package generate-grpc-code-from-protos foo.proto