Allow commnad plugin to take a dir as input
Motivation:

The command plugin requires proto files to be passed as separate inputs.
While expanding these with a wildcard is straightforward, doing so
usually requires the user to also set an import path.

Modifications:

- Allow directories to be used as input to the command plugin. Doing so
  also uses the directory as an import path.
- Improve the help text for the command plugin and provide example calls

Result:

Easier to use the command plugin. In many cases users will only need to
specify the directory containing their protos and the output directory.