Add more documentation about code generation (#83)
Motivation:

Users (especially new ones) often have issues with generating code. This
package provide a number of ways to do this but doesn't do a good job of
explaining the pros and cons of each method or how to use each of them.

Modifications:

- Split up the generating stubs doc into four parts: one article for
each method explaining how to use that approach. For the build plugin
and protoc the content was more or less directly moves from the existing
doc.
- Reprurpose the existing generating stubs doc to explain the pros and
cons of each approach to make it easier for users to decide.

Result:

Better docs